Update email system
