+++ CMakeLists.txt
@@ -8,6 +8,6 @@ include(cmake/TBBGet.cmake)
 tbb_get(TBB_ROOT tbb_root CONFIG_DIR TBB_DIR)
 find_package(TBB)
 
-add_subdirectory(client)
-add_subdirectory(common)
-add_subdirectory(server)
+add_subdirectory(EmailClientServer/client)
+add_subdirectory(EmailClientServer/common)
+add_subdirectory(EmailClientServer/server)
