+++ EmailClientServer/common/include/lib/ctpl.h
