+++ EmailClientServer/common/include/serialization/size_collector.h
