+++ EmailClientServer/server/src/main.cpp
@@ -1,0 +1,9 @@
+#include "Server.h"
+
+#define PORT 4567
+
+int main(int argc, char const *argv[]) {
+    Server server(PORT);
+    server.run_server();
+}
+
