+++ README.md
@@ -1,83 +1,2 @@
 # Networks Lab 2019 (Higher School of Economics)
 
-## Общие замечания
-
-- тип запроса и статус ответа -- 1 байт (`uint8`)
-- все числа в 4 байта -- `uint32` в _little endian_
-- все строки в кодировке UTF_8
-
-### Заголовок всех запросов от клиента серверу
-
-- 4 байта -- размер сообщения
-- 4 байта -- размер адреса
-- строка -- адрес отправителя
-- 1 байт -- тип запроса
-
-### Заголовок всех запросов от сервера клиентам
-
-- 4 байта -- размер сообщения
-
-### Статус ответа
-
-В каждом ответе сервер присылает 1 байт (`uint8`) -- статус ответа на запрос
-Статус 0 -- запрос успешен
-Статус не ноль -- что-то пошло не так (см. ниже тело отчёта)
-
-- _только при статусе отличном от 0_
-- 4 байта -- размер сообщения об ошибке
-- сообщение об ошибке
-
-### 1. A посылает сообщение для B
-
-Запрос:
-- заголовок (тип сообщения 0)
-- 4 байта -- размер адреса отправителя
-- адрес получателя
-- 4 байта -- размер адреса получателя
-- адрес получателя
-- 4 байта -- размер темы письма
-- тема письма
-- 4 байта -- размер тела письма
-- тело письма
-
-Ответ:
-- заголовок
-- 1 байт -- статус запроса
-- _сообщение при статусе != 0_
-
-### 2. A проверяет свои входящие сообщения
-
-Запрос:
-- заголовок (тип сообщения 1)
-
-Ответ:
-- заголовок
-- 1 байт -- статус запроса
-- _тело ниже только при статусе 0_
-  - 4 байта -- количество сообщений (**n**)
-  - _дальше **n** записей_ 
-  - 4 байта -- id письма
-  - 4 байта -- размер адреса отправителя
-  - адрес получателя
-  - 4 байта -- размер темы письма
-  - тема письма
-- _сообщение при статусе != 0_
-
-### 3. A хочет получить сообщение с сервера
-
-Запрос:
-- Заголовок (тип сообщения 2)
-- 4 байта -- id письма
-
-Ответ:
-- 1 байт -- статус запроса
-- _тело ниже только при статусе 0_
-  - 4 байта -- размер адреса отправителя
-  - адрес получателя
-  - 4 байта -- размер адреса получателя
-  - адрес получателя
-  - 4 байта -- размер темы письма
-  - тема письма
-  - 4 байта -- размер тела письма
-  - тело письма
-- _сообщение при статусе != 0_
