--- client/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_executable(client src/main.cpp)
-target_link_libraries(client ${TBB_IMPORTED_TARGETS})
-target_link_libraries(client common)
