--- tcp_template/client_linux/CMakeLists.txt
@@ -1,7 +1,0 @@
-cmake_minimum_required(VERSION 2.8)
-project(client_linux)
-
-set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wall -Wextra")
-
-set(SOURCE_FILES main.c)
-add_executable(client_linux ${SOURCE_FILES})
